增加新api：/user/:user_id，待测试

+++ internal/model/dbmodel.go
@@ -46,7 +46,7 @@ type User struct {
 	MailCode    string    `gorm:"default:null"`
 	Address     string    `gorm:"default:null"`
 	Rating      float32   `gorm:"default:0"`
-	RatingCount float32   `gorm:"default:0"`
+	RatingCount uint   `gorm:"default:0"`
 	Bio         string    `gorm:"default:null"`
 	Token       string
 }
